Got communication working between React and Electron !!!!!
I had to expose some APIs in preload.js for sending and receiving messages between React and Electron.  These don't work in the browser react app.  However, if you run that app (npm start) and then run the electron app (npm run electron-run), the APIs work correctly in the electron app.  (The button uses window.ElectronAPI.send() to send a message to Electron, which is received and printed on the console.